added recieve files to tasks & beat schedule(fixed path - added hierarchy)

+++ users/tasks.py
@@ -211,11 +211,20 @@ def check_robo7_complete():
             #for filename in os.listdir(path=path):
             
                 if os.path.exists(file_to_open_ok)!=-1:
+                    if os.path.exists(file_to_open_dat)!=-1:
+                        try:
+                            os.remove(file_to_open_dat)
+                        except:
+                            pass
 
+                    if os.path.exists(file_to_open_def)!=-1:
+                        try:
+                            os.remove(file_to_open_def)
+                        except:
+                            pass
+                        
                     try:    
                         os.remove(file_to_open_ok)
-                        os.remove(file_to_open_dat)
-                        os.remove(file_to_open_def)
 
                         i.is_complete=True
                         i.save()
